+++ Backup/CommandLineArgs.cs
@@ -39,7 +39,8 @@ public class CommandLineArgs
             "Backup will only create 0 byte placeholder files instead of downloading them, will still create full file structure.")]
     public bool DryRun { get; set; }
 
-    [Option(Required = false, Default = 1, HelpText = "Number of backups to to maintain.")]
+    [Option(Required = false, Default = 1, 
+        HelpText = "Number of previous backups to maintain (in addition to the current backup). Minimum 1 is required for incremental backup to work.")]
     public int BackupsToRotate { get; set; }
 
     [Option("projectstobackup", Required = false, Separator = ',',